fix: Fetch trees from firebase
