+++ lib/screens/tab_views/trees/tree_spinner.dart
@@ -1,45 +1,18 @@
-import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/material.dart';
 
 
+
 class TreeSpinner extends StatefulWidget {
-  TreeSpinner({Key key}) : super(key: key);
+
+  final List<String> treeNamesList;
+  TreeSpinner(this.treeNamesList, {Key key}) : super(key: key);
+  static String dropdownValue = "papaya";
   @override
   _TreeSpinnerState createState() => _TreeSpinnerState();
-  static String dropdownValue = 'Mango';
 
 }
 
-class _TreeSpinnerState extends State<TreeSpinner>{
-
-  //var treesList = ['Mango', 'Papaya', 'Guava', 'Coconut','Banyan','Banana'];
-  List<String> _treeNamesList=[];
-
-  @override
-  void initState() {
-    super.initState();
-    _treeNames();
-  }
-
-  Future _treeNames() async{
-    final ref =  FirebaseDatabase.instance.reference().child("trees");
-    List<String> treeNamesList = [];
-    ref.once().then((DataSnapshot snapshot){
-      Map<dynamic, dynamic> values = snapshot.value;
-      var keys = values.keys;
-      for (var key in keys) {
-        treeNamesList.add(key.toString());
-      }
-      print(treeNamesList);
-    });
-
-    setState(() {
-  _treeNamesList = treeNamesList;
-      print("1");
-      print(_treeNamesList);
-    });
-  }
-
+class _TreeSpinnerState extends State<TreeSpinner> {
 
 
   @override
@@ -62,7 +35,7 @@ class _TreeSpinnerState extends State<TreeSpinner>{
             TreeSpinner.dropdownValue = newValue;
           });
         },
-        items:  _treeNamesList
+        items: widget.treeNamesList
             .map<DropdownMenuItem<String>>((String value) {
           return DropdownMenuItem<String>(
             value: value,
@@ -72,8 +45,4 @@ class _TreeSpinnerState extends State<TreeSpinner>{
       ),
     );
   }
-
-
-
-
 }
